+++ changelog.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - `move_node` bails out early if `new_origin == old_origin` (`d84ff7e`)
 - Coalesce pointer mover events down to `requestAnimationFrame` frame rate (`4b22bc3`)
-- Move Demo Scope documentation to its own `.md` file (``)
+- Move Demo Scope documentation to its own `.md` file (`87845b0`)
 
 ## Fixed
 
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Remove unnecessary floating-point calculations from `elbow_path_into` (`e900283`)
 - Remove unnecessary heap allocations during hex/binary value formatting in `model::content::format_value()` (`f1a1de5`)
 - Remove synchronous text measurement per data node cell construction (`990f34e`)
+- Remove `Vec<String>` allocation from `draw_operator_box` (``)
 
 # [Released]